fix(scout-for-lol): deduplicate betting sweep roster helpers
Why
The architecture refactor exposed duplicated roster helpers and failed the repository duplication ratchet.

What
Move the shared helpers into sweep-roster and remove the now-stale duplication baseline entries.

Verification
- bunx turbo run typecheck test lint --filter=@scout-for-lol/backend
- bun run jscpd
